+++ src/Symfony/Component/Lock/Store/PdoStore.php
@@ -19,7 +19,7 @@ use Symfony\Component\Lock\Exception\InvalidTtlException;
 use Symfony\Component\Lock\Exception\LockConflictedException;
 use Symfony\Component\Lock\Exception\NotSupportedException;
 use Symfony\Component\Lock\Key;
-use Symfony\Component\Lock\StoreInterface;
+use Symfony\Component\Lock\PersistingStoreInterface;
 
 /**
  * PdoStore is a PersistingStoreInterface implementation using a PDO connection.
@@ -34,7 +34,7 @@ use Symfony\Component\Lock\StoreInterface;
  *
  * @author Jérémy Derussé <jeremy@derusse.com>
  */
-class PdoStore implements StoreInterface
+class PdoStore implements PersistingStoreInterface
 {
     use ExpiringStoreTrait;